Move `timestamps` to `committed metadata`
