+++ node/src/utils/test_utils.rs
@@ -5,7 +5,7 @@ use crate::bead::Bead;
 #[cfg(test)]
 use crate::committed_metadata::CommittedMetadata;
 #[cfg(test)]
-use crate::uncommitted_metadata::TimeVec;
+use crate::committed_metadata::TimeVec;
 #[cfg(test)]
 use crate::uncommitted_metadata::UnCommittedMetadata;
 #[cfg(test)]
@@ -27,7 +27,6 @@ pub mod test_utility_functions {
         extra_nonce: i32,
         broadcast_timestamp: Option<Time>,
         signature: Option<Signature>,
-        parent_bead_timestamps: Option<TimeVec>,
     }
 
     #[cfg(test)]
@@ -37,7 +36,6 @@ pub mod test_utility_functions {
                 extra_nonce: 0,
                 broadcast_timestamp: None,
                 signature: None,
-                parent_bead_timestamps: None,
             }
         }
 
@@ -56,11 +54,6 @@ pub mod test_utility_functions {
             self
         }
 
-        pub fn parent_bead_timestamps(mut self, times: TimeVec) -> Self {
-            self.parent_bead_timestamps = Some(times);
-            self
-        }
-
         pub fn build(self) -> UnCommittedMetadata {
             UnCommittedMetadata {
                 extra_nonce: self.extra_nonce,
@@ -68,9 +61,6 @@ pub mod test_utility_functions {
                     .broadcast_timestamp
                     .expect("broadcast_timestamp is required"),
                 signature: self.signature.expect("signature is required"),
-                parent_bead_timestamps: self
-                    .parent_bead_timestamps
-                    .expect("parent_bead_timestamps is required"),
             }
         }
     }
@@ -79,6 +69,7 @@ pub mod test_utility_functions {
         transaction_cnt: u32,
         transactions: Vec<Transaction>,
         parents: std::collections::HashSet<BeadHash>,
+        parent_bead_timestamps: Option<TimeVec>,
         payout_address: Option<P2P_Address>,
         start_timestamp: Option<Time>,
         comm_pub_key: Option<PublicKey>,
@@ -94,6 +85,7 @@ pub mod test_utility_functions {
                 transaction_cnt: 0,
                 transactions: Vec::new(),
                 parents: HashSet::new(),
+                parent_bead_timestamps: None,
                 payout_address: None,
                 start_timestamp: None,
                 comm_pub_key: None,
@@ -115,6 +107,11 @@ pub mod test_utility_functions {
 
         pub fn parents(mut self, parents: HashSet<BeadHash>) -> Self {
             self.parents = parents;
+            self
+        }
+
+        pub fn parent_bead_timestamps(mut self, times: TimeVec) -> Self {
+            self.parent_bead_timestamps = Some(times);
             self
         }
 
@@ -150,6 +147,9 @@ pub mod test_utility_functions {
                 transaction_cnt: self.transaction_cnt,
                 transactions: self.transactions,
                 parents: self.parents,
+                parent_bead_timestamps: self
+                    .parent_bead_timestamps
+                    .expect("parent_bead_timestamps is required"),
                 payout_address: self.payout_address.expect("payout_address is required"),
                 start_timestamp: self
                     .start_timestamp
